Add UI components, disable reconfig testing, fix forward decls.
Configuration and especially reconfiguration needs a lot more work, the
current method for testing is causing problems since all test controllers
are loaded at program start (including their resources) and the
reconfiguration invalidates application window and renderer, causing
loaded textures and surfaces to break.

This could be fixed by simply rewriting how test controllers are queued,
but it highlights a much bigger problem with the current configuration
solution; it would be difficult to implement an in-game options menu
(accessible from a pause screen) that allows users to change graphics
options, as that currently would invalidate all loaded textures and
surfaces.

Users would hate if changing resolution or fullscreen/windowed required
a reload of the game resources (or at least, I know I would). One simple
solution is obviously to only allow "out-of-gameplay" configuration, since
reloading textures and surfaces can be more manageable in that context, but
that is probably a very bad solution.